Build/test updates.
Replaced Build.scala with build.sbt.
Added support Scala 2.11 and cross building with 2.10
Updated scalatest version and updated unit tests to work with the new version.
