+++ src/main/scala/scalawebsocket/WebSocket.scala
@@ -18,7 +18,8 @@ package scalawebsocket
 
 import com.ning.http.client.{websocket, AsyncHttpClient}
 import com.ning.http.client.websocket.{WebSocketTextListener, WebSocketByteListener, WebSocketUpgradeHandler}
-import com.typesafe.scalalogging.log4j.Logging
+import com.typesafe.scalalogging.slf4j._
+// import com.typesafe.scalalogging.log4j.Logging
 
 object WebSocket {
   def apply() = {
@@ -36,7 +37,7 @@ object WebSocket {
   *
   * @param client preconfigured instance of the [[com.ning.http.client.AsyncHttpClient]]
   */
-class WebSocket(client: AsyncHttpClient) extends Logging {
+class WebSocket(client: AsyncHttpClient) extends StrictLogging {
   self =>
 
   type OnTextMessageHandler = String => Unit
